aws/etc/packer/tools/python/requirements.txt
pyyaml == 3.10
